Disable zlib-ng in Platform builds workflow
All ARM targets were failing due to compilation issues related to
zlib-ng Meson wrap. The proper course of action would be to investigate
these issues and either make the wrap work, or provide zlib-ng as a
system library. However, this workflow does not have an actual
maintainer or anybody to put in the effort right now. So apply a quick
workaround of just disabling zlib-ng in this workflow.